Make INTERNAL to BRYAN.
After cloning the BRYAN repo, eg to [~/.BRYAN], clone
this to a folder in the root of the BRYAN repo, eg to
[~/.BRYAN/__bin__], and put that folder on the system
PATH, eg via [export PATH=$PATH:~/.BRYAN/__bin__].

The [./bryan.*] and [./bryan:*] scripts will respect
the layout of the BRYAN repo.

Eg: sourcing [bryan:globals] will set [$BRYAN_ROOT] to
the root of the cloned BRYAN repo, and [bryan.zjot] will
call [bh.zjot] respecting the zjot accumulation path
in the cloned BRYAN repo.